requirements.txt
flask==2.3.2
flask-restful==0.3.9
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
google-generativeai